RFC: rename file_helper to parser
