+++ tests/unit_tests/conftest.py
@@ -2,9 +2,9 @@ import os
 
 import pytest
 
-from wkz.file_helper.fit_parser import FITParser
-from wkz.file_helper.gpx_exporter import gpx_header
-from wkz.file_helper.gpx_parser import GPXParser
+from wkz.parser.fit_parser import FITParser
+from wkz.parser.gpx_exporter import gpx_header
+from wkz.parser.gpx_parser import GPXParser
 
 
 @pytest.fixture
