+++ tests/unit_tests/file_helper/test_auto_naming.py
@@ -2,7 +2,7 @@ import datetime
 
 import pytest
 
-from wkz.file_helper.auto_naming import (
+from wkz.parser.auto_naming import (
     _get_coordinate_not_null,
     _get_daytime_name,
     _get_sport_name,
