+++ wkz/api.py
@@ -7,7 +7,7 @@ from rest_framework import status
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 
-from wkz.file_helper.fit_collector import try_to_mount_device
+from wkz.parser.fit_collector import try_to_mount_device
 
 log = logging.getLogger(__name__)
