+++ wkz/parser/gpx_parser.py
@@ -3,8 +3,8 @@ import logging
 import gpxpy
 import gpxpy.gpx
 
-from wkz.file_helper.parser import Parser
 from wkz.gis.geo import get_total_distance_of_trace
+from wkz.parser.parser import Parser
 
 log = logging.getLogger(__name__)
